ci: ignore more broken clippy lints (#1591)

+++ quic/s2n-quic-core/src/recovery/cubic.rs
@@ -740,7 +740,7 @@ impl Cubic {
     //# where beta_cubic is the CUBIC multiplication decrease factor
     #[inline]
     fn w_cubic(&self, t: Duration) -> f32 {
-        C * (t.as_secs_f32() - self.k.as_secs_f32()).powi(3) + self.w_max as f32
+        C * (t.as_secs_f32() - self.k.as_secs_f32()).powi(3) + self.w_max
     }
 
     //= https://www.rfc-editor.org/rfc/rfc8312#section-4.2
